Assertion and Setup/Teardown examples linked
Assertion and Setup/Teardown examples linked with
existing UsingSeleniumTest.java, removed AssertionsTest.java
and SetupAndTeardownTest.java
